mission10 리펙토링
