+++ ecosystem.config.js
@@ -2,8 +2,8 @@ module.exports = {
   apps: [
     {
       name: 'panda-market',
-      script: 'npm',
-      args: 'start',
+      script: './dist/main.js',
+      args: 'run start',
       env: {
         NODE_ENV: 'production',
       },
