+++ package.json
@@ -1,7 +1,7 @@
 {
   "scripts": {
     "start": "node dist/main.js",
-    "dev": "nodemon --watch src src/main.ts",
+    "dev": "cross-env NODE_ENV=development nodemon --watch src src/main.ts",
     "build": "tsc",
     "typecheck": "tsc --noEmit",
     "prisma:generate": "prisma generate",
@@ -22,6 +22,7 @@
     "@types/multer": "^1.4.12",
     "@types/superstruct": "^0.5.0",
     "@types/supertest": "^6.0.3",
+    "cross-env": "^7.0.3",
     "dotenv-cli": "^8.0.0",
     "jest": "^29.7.0",
     "nodemon": "^3.1.9",
